# server/openapi.yaml
openapi: 3.0.3
info:
  title: Metro Verify API
  description: >
    Digital Legal Metrology Verification & Lifecycle Management Platform (SIH26036).
    Covers Owner, Administrator, and LMO/GATC Officer endpoints, plus the AI chatbot.
    Endpoints marked "(stored procedure)" call a MySQL stored procedure for an atomic,
    multi-step operation rather than a plain Prisma query.
  version: "1.0.0"
servers:
  - url: http://localhost:4000
    description: Local development backend
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Instrument:
      type: object
      properties:
        instrumentId: { type: string, example: "INST-TS-08" }
        type: { type: string, example: "Non-Automatic Weighing Instrument" }
        manufacturer: { type: string }
        model: { type: string }
        serialNumber: { type: string }
        ownerId: { type: string }
        status:
          type: string
          enum: [REGISTERED, SCHEDULED, ASSIGNED, IN_PROGRESS, VERIFIED, EXPIRED]
...
security:
  - bearerAuth: []
paths:
...
  /api/instruments:
    get:
      tags: [Owner - Instruments]
      summary: List the logged-in owner's instruments
      responses:
        "200":
          description: List of instruments
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Instrument" }
    post:
      tags: [Owner - Instruments]
      summary: Register a new instrument
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Instrument" }
      responses:
        "201": { description: Instrument registered, content: { application/json: { schema: { $ref: "#/components/schemas/Instrument" } } } }
  /api/instruments/{id}:
    get:
      tags: [Owner - Instruments]
      summary: Get a single instrument
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: Instrument detail, content: { application/json: { schema: { $ref: "#/components/schemas/Instrument" } } } }
  /api/instruments/{id}/passport:
    get:
      tags: [Owner - Instruments]
      summary: Get the Digital Instrument Passport (full lifecycle record — distinct from a Certificate)
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: Passport with lifecycle timeline events }

  # ---------------- OWNER: CERTIFICATES ----------------
...
  /api/lmo/batches/{id}:
    get:
      tags: [LMO - Assignments]
      summary: Get a batch and its instrument-level progress, scoped to the logged-in officer's portion
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: Batch detail with per-instrument status }
  /api/lmo/instruments/{id}:
    get:
      tags: [LMO - Assignments]
      summary: Get instrument detail for a specific assignment (only if assigned to the logged-in officer)
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: Instrument detail, content: { application/json: { schema: { $ref: "#/components/schemas/Instrument" } } } }
        "403": { description: Not assigned to this officer }

  # ---------------- LMO: INSPECTIONS ----------------
...